Move XML generation to ConstantMetaData
Also activate the formatOutput option on the XMLDocument.
One problem with it however is that whitespace sensitivity means we still need to handle the <term> tag manually.
The other problem is that the indentation used by ext/dom is 2 spaces, so cater to this until I fix this in ext/dom.